Fixed pid regulator

+++ Node_2/pid.h
@@ -16,9 +16,9 @@ typedef struct Pid_s{
 	double p_factor;
 	double i_factor;
 	double d_factor;
-	int8_t error;
-	int8_t sum_error;
-	int8_t last_error;
+	int16_t error;
+	int16_t sum_error;
+	int16_t last_error;
 	int u;
 	int8_t ref;
 	int8_t measured;
